PID climate heat and cool only
When a PID climate supports both heating and cooling, the user may select mode HEAT_COOL as well as HEAT (only) and COOL (only)

It is unexpected if the controlled output still heats and cools when one of the modes HEAT or COOL and not HEAT_COOL is selected.

For a use case, I have a brew temp controller that I sometimes want to run in heat only mode. The reason is two fold, because the cooling isn't critical (bottle carb phase) and in my setup the cooling requires refreshing an ice reservoir. If I'm not refreshing that reservoir because I don't care about cooling during this time, then the cool output does nothing useful (but does make noise). So hence I want to run it in heat only. In this case when the temperature rises above setpoint, I want the control to do nothing - as if the device was only capable of heating.

Additionally (not in this change proposal) it may be well to provide a configuration option to add the heat only and cool only modes or to not add them, so they are not selectable from the interface. The latter may even be the default.